+++ src/ent/bai_result.h
@@ -1,6 +1,8 @@
 #ifndef BAI_RESULT_H
 #define BAI_RESULT_H
 
+#include <stdint.h>
+
 typedef enum {
   BAI_RESULT_STATUS_NONE,
   BAI_RESULT_STATUS_MAX_ITERATIONS,
@@ -12,7 +14,7 @@ typedef enum {
 
 typedef struct BAIResult BAIResult;
 
-void bai_result_reset(BAIResult *bai_result, int time_limit_seconds);
+void bai_result_reset(BAIResult *bai_result, uint64_t time_limit_seconds);
 BAIResult *bai_result_create(void);
 void bai_result_destroy(BAIResult *bai_result);
 void bai_result_set_best_arm(BAIResult *bai_result, int best_arm);
